requirements.txt
Flask==3.0.3
matplotlib
numpy
opencv-contrib-python
opencv-python
pandas
requests
gunicorn
ultralytics
torch
transformers
torchvision
pillow
